Added Movement
Added a basic movement system using the actual keycodes for movement. I am not a big fan of this style of key handling, however this DOES work and it DOES allow for animation.

Something to note is that, apparently, the Arrow keycodes are as-is VERY BROKEN in p5. Stick to the proper keys.